Move 'identify_policies.py' to 'hexarena.scripts'
